requirements.txt
web3>=6.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
# Optional: TensorFlow for LSTM predictor (not required for core functionality)
# tensorflow>=2.15.0
